+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### 0.0.4
+
+**Commit Delta**: [Change from 0.0.3 release](https://github.com/plus3it/tardigrade/compare/0.0.3...0.0.4)
+
+**Released**: 2020.01.08
+
+**Summary**:
+
+*   Add metric filters
+*   Remove unused variables
+*   Add metric alarms
+*   Remove AWSSupportAccess policy from user and add it to a group instead
+
 ### 0.0.3
 
 **Commit Delta**: [Change from 0.0.2 release](https://github.com/plus3it/tardigrade/compare/0.0.2...0.0.3)